+++ changelog.md
@@ -2,7 +2,7 @@
 ## 4.9.0
 * Fix regression from 4.8.1 that would generate wrong copy methods for object arrays
 * extracted AzFuncTestHelper to a separate repository
-* Add CaffoaClientStringException and CAffoaClientJsonException\<T> for easier error error handling in servers
+* Add CaffoaClientStringException and CaffoaClientJsonException\<T> for easier error handling in servers
 * allow to create objects with anyOf
 * allow to create array objects
 * handle query parameters that have characters that are not allowed in C# identifiers (e.g. types[])